# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/hello":{"get":{"summary":"Root","operationId":"root_hello_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/auth/hello":{"get":{"tags":["auth"],"summary":"Hello","operationId":"hello_auth_hello_get","responses":{"200":{"description":"To test if the route is alive and working well","content":{"application/json":{"schema":{}}}}}}},"/auth/login":{"post":{"tags":["auth"],"summary":"Login Route","operationId":"login_route_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_route_auth_login_post"}}},"required":true},"responses":{"202":{"description":"Successful login","content":{"application/json":{"schema":{},"example":{"access_token":"your_token","token_type":"bearer"}}}},"404":{"description":"User not found.","content":{"application/json":{"example":{"detail":"User not found."}}}},"401":{"description":"Password is incorrect. Try again","content":{"application/json":{"example":{"detail":"Password is incorrect. Try again"}}}},"500":{"description":"Internal server error","content":{"application/json":{"example":{"detail":"An unexpected error occurred: (e)"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/register":{"post":{"tags":["auth"],"summary":"Register Route","operationId":"register_route_auth_register_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_register_route_auth_register_post"}}},"required":true},"responses":{"201":{"description":"User registered successfully","content":{"application/json":{"schema":{},"example":{"message":"User account successfully created.","data":{"username":"johndoe","email":"johndoe@example.com","full_name":"John Doe"}}}}},"409":{"description":"User with this username already exists","content":{"application/json":{"example":{"message":"User with this username already exists"}}}},"500":{"description":"An unexpected error occurred during user registration. Please try again later.","content":{"application/json":{"example":{"detail":"An unexpected error occurred during user registration. Please try again later."}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/logout":{"post":{"tags":["auth"],"summary":"Logout User","operationId":"logout_user_auth_logout_post","responses":{"204":{"description":"Logout user"}}}},"/user/hello":{"get":{"tags":["user"],"summary":"Hello","operationId":"hello_user_hello_get","responses":{"200":{"description":"To test if the route is alive and working well","content":{"application/json":{"schema":{}}}}}}},"/user/me":{"get":{"tags":["user"],"summary":"Get Curr User Route","operationId":"get_curr_user_route_user_me_get","responses":{"200":{"description":"Current logged in user details","content":{"application/json":{"schema":{},"example":{"username":"johndoe","email":"johndoe@example.com","full_name":"John Doe"}}}},"401":{"description":"Could not validate credentials. Try logging in again","content":{"application/json":{"example":{"detail":"Could not validate credentials. Try logging in again"}}}},"404":{"description":"User not found.","content":{"application/json":{"example":{"detail":"User not found."}}}}},"security":[{"OAuth2PasswordBearer":[]}]},"delete":{"tags":["user"],"summary":"Delete User Route","operationId":"delete_user_route_user_me_delete","responses":{"200":{"description":"Delete user","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/user/file":{"get":{"tags":["user"],"summary":"Get File Details Route","operationId":"get_file_details_route_user_file_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"filename","in":"query","required":true,"schema":{"type":"string","title":"Filename"}}],"responses":{"200":{"description":"Successful file retrieval","content":{"application/json":{"schema":{},"example":{"name":"example.txt","uploaded_at":"2024-07-14T12:34:56","tokenized":true,"embedded":false,"processed":false}}}},"404":{"description":"User not found.","content":{"application/json":{"example":{"detail":"User not found."}}}},"500":{"description":"Internal server error","content":{"application/json":{"example":{"detail":"An unexpected error occurred: (e)"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/user/file/list":{"get":{"tags":["user"],"summary":"List User Files Route","operationId":"list_user_files_route_user_file_list_get","responses":{"200":{"description":"Files successfully retrieves from user files","content":{"application/json":{"schema":{},"example":[{"name":"example.txt","uploaded_at":"2024-07-14T12:34:56","tokenized":true,"embedded":false,"processed":false}]}}},"404":{"description":"User not found.","content":{"application/json":{"example":{"detail":"User not found."}}}},"500":{"description":"Internal server error","content":{"application/json":{"example":{"detail":"An unexpected error occurred: (e)"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/docu/hello":{"get":{"tags":["docu"],"summary":"Hello","operationId":"hello_docu_hello_get","responses":{"200":{"description":"To test if the route is alive and working well","content":{"application/json":{"schema":{}}}}}}},"/docu/get-tokens":{"get":{"tags":["docu"],"summary":"Get Tokens Route","operationId":"get_tokens_route_docu_get_tokens_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"filename","in":"query","required":true,"schema":{"type":"string","title":"Filename"}}],"responses":{"200":{"description":"Tokens retrieved successfully","content":{"application/json":{"schema":{},"example":{"message":"Token 'filename' has successfully been retrieved.","data":{"processed":"2023-01-01T00:00:00","doc_loader_used":"loader_name","token_count":100,"chunk_count":10,"chunks":[{"key":"value"}]}}}}},"404":{"description":"User not found or file not found in the database","content":{"application/json":{"example":{"detail":"User not found."}}}},"409":{"description":"File '{file}' has not yet been tokenized.","content":{"application/json":{"example":{"detail":"File 'filename' has not yet been tokenized."}}}},"500":{"description":"An unexpected error occurred","content":{"application/json":{"example":{"detail":"An unexpected error occurred"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/docu/get-tscc":{"get":{"tags":["docu"],"summary":"Get Tscc Route","operationId":"get_tscc_route_docu_get_tscc_get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"filename","in":"query","required":true,"schema":{"type":"string","title":"Filename"}}],"responses":{"200":{"description":"TSCC data retrieved successfully","content":{"application/json":{"schema":{},"example":{"message":"Token 'filename' has successfully been retrieved.","data":{"processed":"2023-01-01T00:00:00","process_time":0.5,"model_used":"model_name","token_count":100,"chunk_count":10,"chunks":["chunk1","chunk2"]}}}}},"404":{"description":"User not found or file not found in the database","content":{"application/json":{"example":{"detail":"User not found."}}}},"409":{"description":"File '{file}' is not yet processed for TSCC.","content":{"application/json":{"example":{"detail":"File 'filename' is not yet processed for TSCC."}}}},"500":{"description":"An unexpected error occurred","content":{"application/json":{"example":{"detail":"An unexpected error occurred"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/docu/upload":{"post":{"tags":["docu"],"summary":"Upload File Route","operationId":"upload_file_route_docu_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_file_route_docu_upload_post"}}},"required":true},"responses":{"200":{"description":"Upload a file","content":{"application/json":{"schema":{}}}},"201":{"description":"File uploaded successfully","content":{"application/json":{"example":{"message":"File 'filename' successfully uploaded."}}}},"404":{"description":"User not found.","content":{"application/json":{"example":{"detail":"User not found."}}}},"409":{"description":"File '{file}' already exists in the server.","content":{"application/json":{"example":{"detail":"File 'filename' already exists in the server."}}}},"500":{"description":"An unexpected error occurred during file upload","content":{"application/json":{"example":{"detail":"An unexpected error occurred during file upload. Please try again later."}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/docu/gen-tokens":{"post":{"tags":["docu"],"summary":"Generate Tokens Route","operationId":"generate_tokens_route_docu_gen_tokens_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_generate_tokens_route_docu_gen_tokens_post"}}},"required":true},"responses":{"200":{"description":"Tokens generated successfully","content":{"application/json":{"schema":{},"example":{"message":"File 'filename' has been tokenized and embedded in ChromaDB successfully.","data":{"tokens":"example_tokens_data"}}}}},"404":{"description":"The file '{file}' does not exist on the server.","content":{"application/json":{"example":{"detail":"The file 'filename' does not exist on the server."}}}},"409":{"description":"File '{file}' has already been tokenized.","content":{"application/json":{"example":{"detail":"File 'filename' has already been tokenized."}}}},"400":{"description":"Tokenization failed for file '{file}': {error}","content":{"application/json":{"example":{"detail":"Tokenization failed for file 'filename': error_message"}}}},"500":{"description":"An unexpected error occurred during token generation","content":{"application/json":{"example":{"detail":"An unexpected error occurred during token generation. Please try again later."}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/docu/process-tscc":{"post":{"tags":["docu"],"summary":"Process Tscc Route","operationId":"process_tscc_route_docu_process_tscc_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_process_tscc_route_docu_process_tscc_post"}}},"required":true},"responses":{"200":{"description":"Process TSCC for a file","content":{"application/json":{"schema":{}}}},"202":{"description":"TSCC processing initiated in the background","content":{"application/json":{"example":{"message":"File 'filename' is currently being processed for TSCC."}}}},"404":{"description":"The file '{file}' does not exist on the database.","content":{"application/json":{"example":{"detail":"The file 'filename' does not exist on the database."}}}},"409":{"description":"Conflict: File already processed or not tokenized","content":{"application/json":{"example":{"detail":"File 'filename' has not yet been tokenized."}}}},"400":{"description":"TSCC processing failed for '{file}'.","content":{"application/json":{"example":{"detail":"TSCC processing failed for 'filename'."}}}},"500":{"description":"Internal server error during TSCC processing","content":{"application/json":{"example":{"detail":"An unexpected error occurred during TSCC processing. Please try again later."}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/docu/query-rag":{"post":{"tags":["docu"],"summary":"Query Rag Route","operationId":"query_rag_route_docu_query_rag_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_query_rag_route_docu_query_rag_post"}}},"required":true},"responses":{"200":{"description":"Query successful with RAG model","content":{"application/json":{"schema":{},"example":{"message":"RAG Query successful","data":{"query":"query_text","response":"response_text"}}}}},"404":{"description":"The file '{file}' does not exist on the server.","content":{"application/json":{"example":{"detail":"The file 'filename' does not exist on the server."}}}},"409":{"description":"Conflict: File not tokenized or not yet embedded","content":{"application/json":{"example":{"detail":"File 'filename' has not yet been tokenized."}}}},"500":{"description":"Internal server error during query processing","content":{"application/json":{"example":{"detail":"An unexpected error occurred during query processing. Please try again later."}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/docu/":{"delete":{"tags":["docu"],"summary":"Delete File Route","operationId":"delete_file_route_docu__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"filename","in":"query","required":true,"schema":{"type":"string","title":"Filename"}}],"responses":{"200":{"description":"File successfully deleted","content":{"application/json":{"schema":{},"example":{"message":"File 'filename' successfilly deleted."}}}},"400":{"description":"Failed to delete file from database","content":{"application/json":{"example":{"detail":"File 'filename' was not deleted from the database successflly."}}}},"500":{"description":"Internal server error during file deletion","content":{"application/json":{"example":{"detail":"An unexpected error occurred during file deletion. Please try again later."}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/docu/delete-tokens":{"delete":{"tags":["docu"],"summary":"Delete Tokens Route","operationId":"delete_tokens_route_docu_delete_tokens_delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"filename","in":"query","required":true,"schema":{"type":"string","title":"Filename"}}],"responses":{"200":{"description":"Successfully deleted {file}'s tokens from the database.","content":{"application/json":{"schema":{},"example":{"message":"Successfully deleted filename's tokens from the database."}}}},"400":{"description":"Failed to delete document with tokens_id {tokens_id}.","content":{"application/json":{"example":{"detail":"Failed to delete document with tokens_id {tokens_id}."}}}},"404":{"description":"The file '{file}' does not exist on the database.","content":{"application/json":{"example":{"detail":"The file '{file}' does not exist on the database."}}}},"500":{"description":"Internal server error during token deletion","content":{"application/json":{"example":{"detail":"An unexpected error occurred during token deletion. Please try again later."}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/docu/delete-tscc":{"delete":{"tags":["docu"],"summary":"Delete Tscc Route","operationId":"delete_tscc_route_docu_delete_tscc_delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"filename","in":"query","required":true,"schema":{"type":"string","title":"Filename"}}],"responses":{"200":{"description":"Successfully deleted {file}'s tokens from the database.","content":{"application/json":{"schema":{},"example":{"message":"Successfully deleted filename's tokens from the database."}}}},"400":{"description":"File '{file}' is not yet processed for TSCC.","content":{"application/json":{"example":{"detail":"File '{file}' is not yet processed for TSCC."}}}},"404":{"description":"The file '{file}' does not exist on the database.","content":{"application/json":{"example":{"detail":"The file '{file}' does not exist on the database."}}}},"500":{"description":"Internal server error during tscc deletion","content":{"application/json":{"example":{"detail":"An unexpected error occurred during tscc deletion. Please try again later."}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_generate_tokens_route_docu_gen_tokens_post":{"properties":{"filename":{"type":"string","title":"Filename"},"pdf_loader":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pdf Loader"},"overwrite":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Overwrite"}},"type":"object","required":["filename"],"title":"Body_generate_tokens_route_docu_gen_tokens_post"},"Body_login_route_auth_login_post":{"properties":{"remember_me":{"title":"Remember Me"},"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["remember_me","username","password"],"title":"Body_login_route_auth_login_post"},"Body_process_tscc_route_docu_process_tscc_post":{"properties":{"filename":{"type":"string","title":"Filename"},"llm":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Llm"}},"type":"object","required":["filename"],"title":"Body_process_tscc_route_docu_process_tscc_post"},"Body_query_rag_route_docu_query_rag_post":{"properties":{"filename":{"type":"string","title":"Filename"},"query":{"type":"string","title":"Query"}},"type":"object","required":["filename","query"],"title":"Body_query_rag_route_docu_query_rag_post"},"Body_register_route_auth_register_post":{"properties":{"username":{"type":"string","title":"Username"},"full_name":{"type":"string","title":"Full Name"},"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","full_name","email","password"],"title":"Body_register_route_auth_register_post"},"Body_upload_file_route_docu_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_file_route_docu_upload_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"auth/token"}}}}}}